Move projects router to separate file
