+++ apps/server-go-htmx/main.go
@@ -15,9 +15,11 @@ func main() {
 	log.Fatal(server(":5003"))
 }
 
+// server.go
 func server(addr string) error {
 	router := http.NewServeMux()
-	router.Handle("/", projectsRouter())
+	router.Handle("/api/v1/", http.StripPrefix("/api/v1", projectsRouter()))
+	router.HandleFunc("/", NotFoundHandler)
 
 	apiServer := http.Server{
 		Addr:    addr,
@@ -27,39 +29,14 @@ func server(addr string) error {
 	return apiServer.ListenAndServe()
 }
 
-func projectsRouter() *http.ServeMux {
-	router := http.NewServeMux()
-	router.HandleFunc("GET /projects", GetAllProjects)
-	router.HandleFunc("GET /projects/{id}", GetProjectById)
-	router.HandleFunc("POST /projects", CreateProject)
-	router.HandleFunc("PUT /projects/{id}", UpdateProjectById)
-	router.HandleFunc("DELETE /projects/{id}", DeleteProject)
+func NotFoundHandler(w http.ResponseWriter, r *http.Request) {
+	scheme := "http"
+	if r.TLS != nil {
+		scheme = "https"
+	}
+	url := fmt.Sprintf("%s://%s%s", scheme, r.Host, r.URL.Path)
 
-	return router
-}
-
-func GetAllProjects(w http.ResponseWriter, r *http.Request) {
-	json.NewEncoder(w).Encode(ResponseMessage{Msg: "All Projects"})
-}
-
-func GetProjectById(w http.ResponseWriter, r *http.Request) {
-	projectId := r.PathValue("id")
-	json.NewEncoder(w).Encode(ResponseMessage{Msg: "Project with Id " + projectId})
-}
-
-func CreateProject(w http.ResponseWriter, r *http.Request) {
-	w.WriteHeader(http.StatusCreated)
-	json.NewEncoder(w).Encode(ResponseMessage{Msg: "Project created"})
-}
-
-func UpdateProjectById(w http.ResponseWriter, r *http.Request) {
-	projectId := r.PathValue("id")
-	json.NewEncoder(w).Encode(ResponseMessage{
-		Msg: fmt.Sprintf("Project %s updated", projectId)})
-}
-
-func DeleteProject(w http.ResponseWriter, r *http.Request) {
-	projectId := r.PathValue("id")
-	json.NewEncoder(w).Encode(ResponseMessage{
-		Msg: fmt.Sprintf("Project %s deleted", projectId)})
+	w.WriteHeader(http.StatusNotFound)
+	json.NewEncoder(w).Encode(map[string]any{
+		"Msg": "NotFound - " + url})
 }
